+++ src/components/player/chapters/ChapterIndicator.vue
