--- src/styles/_marquee.scss
@@ -1,4 +1,0 @@
-.marquee {
-  transform: translateX(0);
-  animation: stop-and-scroll 15s linear infinite;
-}
